更改接口方法：GetInt64 -> GetInt

+++ reqparse/parser.go
@@ -205,13 +205,15 @@ func (p *RequestParser) Valid(k string, tags []string, v reflect.Value) error {
 func (p *RequestParser) autoSetValue(geter ValueGetter, k string, v reflect.Value) error {
 	switch v.Kind() {
 	case reflect.Int, reflect.Int8, reflect.Int16, reflect.Int32, reflect.Int64:
-		val, err := geter.GetInt64(k)
+		val, err := geter.GetInt(k)
 		if err != nil {
 			return ValueError{k, fmt.Sprintf("'%s' is not a valid choice", geter.GetString(k))}
 		}
 		v.SetInt(val)
+
 	case reflect.String:
 		v.SetString(geter.GetString(k))
+
 	case reflect.Float32, reflect.Float64:
 	case reflect.Bool:
