+++ reqparse/value.go
@@ -5,7 +5,7 @@ import "github.com/astaxie/beego"
 type ValueGetter interface {
 	GetString(key string, def ...string) string
 	GetStrings(key string, def ...[]string) []string
-	GetInt64(key string, def ...int64) (val int64, err error)
+	GetInt(key string, def ...int64) (val int64, err error)
 	GetBool(key string, def ...bool) (val bool, err error)
 }
 
@@ -21,7 +21,7 @@ func (v *FromValues) GetStrings(k string, def ...[]string) []string {
 	return v.ctl.GetStrings(k)
 }
 
-func (v *FromValues) GetInt64(key string, def ...int64) (val int64, err error) {
+func (v *FromValues) GetInt(key string, def ...int64) (val int64, err error) {
 	return v.ctl.GetInt64(key)
 }
